+++ tests/trivial.hpp
@@ -42,9 +42,9 @@ namespace tBLAS_test
     std::vector<std::vector<T>> Trivial::transpose(const std::vector<std::vector<T>> &A)
     {
         std::vector<std::vector<T>> B(A[0].size(), std::vector<T>(A.size()));
-        for (int i = 0; i < A.size(); ++i)
+        for (size_t i = 0; i < A.size(); ++i)
         {
-            for (int j = 0; j < A[0].size(); ++j)
+            for (size_t j = 0; j < A[0].size(); ++j)
             {
                 B[j][i] = A[i][j];
             }
@@ -56,11 +56,11 @@ namespace tBLAS_test
     std::vector<std::vector<T>> Trivial::matmul(const std::vector<std::vector<T>> &A, const std::vector<std::vector<T>> &B)
     {
         std::vector<std::vector<T>> C(A.size(), std::vector<T>(B[0].size()));
-        for (int i = 0; i < A.size(); ++i)
+        for (size_t i = 0; i < A.size(); ++i)
         {
-            for (int j = 0; j < B[0].size(); ++j)
+            for (size_t j = 0; j < B[0].size(); ++j)
             {
-                for (int k = 0; k < B.size(); ++k)
+                for (size_t k = 0; k < B.size(); ++k)
                 {
                     C[i][j] += A[i][k] * B[k][j];
                 }
